exercise 2
When you finish Exercise 1, navigate to the directory above the my_folder directory and delete all the content you generated in exercise 1 with one command.

WARNING: The rm method is DANGEROUS. Once you run this command, there is no undoing it. It is permanent. Make sure you are in the correct directory and that you use the correct name of the directory you want to delete.