exercise 5
Identify the Constructors for each of the following methods and classify each method as either a "Static" or a "Prototype" method:

substring
create
fromCharCode
slice
toString
It might be that more than one constructor provides a method with a given name.Make sure you list them all in your answer. You can limit your search for methods to the String, Object, Array, and Number constructors.